Rework navigation to accept arguments
And use those to name the edit view (previously known as the add view)
