+++ app/TimeKeep.App/Views/EditEntryView.axaml
@@ -1,0 +1,111 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:c="using:TimeKeep.App.Controls"
+             xmlns:converters="using:TimeKeep.App.Converters"
+             xmlns:vm="using:TimeKeep.App.ViewModels"
+             xmlns:fi="using:FluentAvalonia.FluentIcons"
+             mc:Ignorable="d" d:DesignWidth="320" d:DesignHeight="600"
+             x:Class="TimeKeep.App.Views.EditEntryView"
+             x:DataType="vm:EditEntryViewModel">
+    <UserControl.Styles>
+		<Style Selector="Label">
+			<Setter Property="Margin" Value="0,10,0,5"/>
+		</Style>
+		<Style Selector="AutoCompleteBox ListBoxItem">
+			<Setter Property="Padding" Value="0"/>
+		</Style>
+    </UserControl.Styles>
+	<UserControl.Resources>
+		<converters:LocationConverter x:Key="LocationConverter"/>
+		<Flyout x:Key="ErrorFlyout" Placement="Center">
+			<StackPanel>
+				<TextBlock Classes="h2" Text="Error"/>
+				<TextBlock AttachedToVisualTree="OnErrorMessageBlockAttachedToVisualTree"
+				           TextWrapping="WrapWithOverflow"/>
+				<Button Click="OnErrorFyloutOK" Content="OK"
+				        Margin="0,5,0,0" HorizontalAlignment="Right"/>
+			</StackPanel>
+		</Flyout>
+	</UserControl.Resources>
+
+	<Grid RowDefinitions="Auto,*,Auto">
+	    <TextBlock Grid.Row="0"
+	               Classes="h1"
+                   Margin="20,10,20,0"
+	               Text="{Binding Title}"/>
+
+		<ScrollViewer Grid.Row="1" Padding="20,0,20,10">
+			<StackPanel>
+				<Label Content="Start" Target="StartPicker"/>
+				<c:DateTimePicker x:Name="StartPicker" DateTime="{Binding StartTime}"/>
+
+				<StackPanel Orientation="Horizontal" Spacing="20" Margin="0,10,0,5">
+					<Label Content="End" Target="EndPicker" Margin="0" VerticalAlignment="Center"/>
+					<CheckBox IsChecked="{Binding EndTime, Converter={x:Static ObjectConverters.IsNotNull}, Mode=OneWay}"
+					          IsCheckedChanged="OnEndTimeIsCheckedChanged"/>
+				</StackPanel>
+				<c:DateTimePicker x:Name="EndPicker"
+				                  IsEnabled="{Binding EndTime, Converter={x:Static ObjectConverters.IsNotNull}}"
+				                  DateTime="{Binding EndTime}"/>
+
+				<Label Content="Project" Target="ProjectBox"/>
+				<ComboBox x:Name="ProjectBox"
+				          SelectedItem="{Binding SelectedProject}"
+				          ItemsSource="{Binding Projects}">
+					<ComboBox.ItemTemplate>
+						<DataTemplate>
+							<TextBlock Text="{Binding Name}"/>
+						</DataTemplate>
+					</ComboBox.ItemTemplate>
+				</ComboBox>
+
+				<Label Content="Categories" Target="CategoriesBox"/>
+				<c:CategoriesDisplay Categories="{Binding Categories}" CanRemove="True"/>
+				<Panel>
+					<AutoCompleteBox x:Name="CategoriesBox" ItemsSource="{Binding ExistingCategories}"
+					                 KeyDown="OnAddCategoryKeyDown">
+						<AutoCompleteBox.ItemTemplate>
+							<DataTemplate DataType="x:String">
+								<Grid Background="Transparent" PointerPressed="OnCategoryPressed" Tag="{Binding}">
+									<TextBlock Text="{Binding}" Margin="12,9"/>
+								</Grid>
+							</DataTemplate>
+						</AutoCompleteBox.ItemTemplate>
+					</AutoCompleteBox>
+					<Button CornerRadius="0,3,3,0"
+					        Height="{Binding #CategoriesBox.Bounds.Height}"
+					        HorizontalAlignment="Right"
+					        KeyboardNavigation.IsTabStop="False"
+					        Click="OnAddCategoryClicked">
+						<fi:FluentIcon Icon="Add20Regular"/>
+					</Button>
+				</Panel>
+
+				<Label Content="Location" Target="LocationBox"/>
+				<ComboBox x:Name="LocationBox"
+				          SelectedItem="{Binding SelectedLocation}"
+				          ItemsSource="{Binding Locations}">
+					<ComboBox.ItemTemplate>
+						<DataTemplate>
+							<TextBlock Text="{Binding Converter={StaticResource LocationConverter}, FallbackValue='Empty'}"/>
+						</DataTemplate>
+					</ComboBox.ItemTemplate>
+				</ComboBox>
+			</StackPanel>
+		</ScrollViewer>
+
+		<StackPanel Grid.Row="2"
+		            Orientation="Horizontal"
+		            HorizontalAlignment="Right"
+		            Margin="20,10,20,10"
+		            Spacing="10">
+			<Button Classes="accent"
+			        Content="{Binding SaveVerb}"
+			        Click="OnAddClick"/>
+			<Button Content="Cancel"
+			        Click="OnCancelClick"/>
+		</StackPanel>
+    </Grid>
+</UserControl>
